+++ princeofversions/src/androidMain/kotlin/com/infinum/princeofversions/AndroidApplicationVersionProvider.kt
@@ -7,11 +7,11 @@ import android.os.Build
 /**
  * Provides the current version of the application.
  */
-public typealias ApplicationVersionProvider = BaseApplicationVersionProvider<Int>
+public typealias ApplicationVersionProvider = BaseApplicationVersionProvider<Long>
 
 internal class AndroidApplicationVersionProvider(context: Context) : ApplicationVersionProvider {
 
-    private val version: Int = try {
+    private val version: Long = try {
         val packageInfo = context.packageManager.getPackageInfo(context.packageName, 0)
         val longVersionCode = if (Build.VERSION.SDK_INT >= Build.VERSION_CODES.P) {
             packageInfo.longVersionCode
@@ -19,11 +19,10 @@ internal class AndroidApplicationVersionProvider(context: Context) : Application
             @Suppress("DEPRECATION")
             packageInfo.versionCode.toLong()
         }
-
-        longVersionCode.toInt()
+        longVersionCode
     } catch (e: PackageManager.NameNotFoundException) {
         throw kotlin.IllegalStateException("Could not find package name", e)
     }
 
-    override fun getVersion(): Int = version
+    override fun getVersion(): Long = version
 }
